refactor(core): split tools/mod.rs into definitions + tests modules (<1000)
tools/mod.rs was 2834 lines. Extract:
- definitions.rs: the static `ToolDef` registry entries + small inline execs.
- tests_a.rs / tests_b.rs: the ~80-test #[cfg(test)] module, split in two.

mod.rs (665 lines) keeps the permission types, registry accessors, tool-type
predicates, and execute_tool dispatch. Every rustyclaw-core file is now under
1000 lines. No behavior change.
